requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
astor==0.8.1
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
azure-core==1.29.5
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
blessed==1.20.0
Brotli==1.1.0
bs4==0.0.1
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.0
Chroma==0.2.0
chroma-hnswlib==0.7.3
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.0
cryptography==41.0.4
cycler==0.12.1
dataclasses-json==0.6.1
distlib==0.3.7
distro==1.8.0
dnspython==2.4.2
ecdsa==0.18.0
faiss-cpu==1.7.4
fastapi==0.103.2
ffmpy==0.3.1
filelock==3.12.4
firebase-admin==6.2.0
flatbuffers==23.5.26
fonttools==4.45.0
frozenlist==1.4.0
fsspec==2023.9.2
ftfy==6.1.1
git-python==1.0.3
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.12.0
google-api-python-client==2.109.0
google-auth==2.23.2
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpcio==1.59.3
grpcio-status==1.59.3
gunicorn==21.2.0
h11==0.14.0
httpcore==0.18.0
httplib2==0.22.0
httptools==0.6.0
httpx==0.25.0
huggingface-hub==0.17.3
humanfriendly==10.0
identify==2.5.30
idna==3.4
imageio==2.31.5
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
importlib-resources==6.1.0
imutils==0.5.4
inquirer==3.1.3
isodate==0.6.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.0.334
langsmith==0.0.63
lazy_loader==0.3
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
mdurl==0.1.2
mongoengine==0.27.0
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.7
msrest==0.7.1
multidict==6.0.4
mutagen==1.47.0
mypy-extensions==1.0.0
networkx==3.1
nodeenv==1.8.0
numpy==1.26.0
oauthlib==3.2.2
onnxruntime==1.16.1
ooba==0.0.21
openai==1.1.2
openai-clip==1.0.1
opencv-python-headless==4.8.1.78
outcome==1.2.0
overrides==7.4.0
packaging==23.2
pdfdocument==4.0.0
pgvector==0.2.3
Pillow==10.0.1
pip-check-reqs==2.5.3
pipdeptree==2.13.1
platformdirs==3.11.0
posthog==3.0.2
pre-commit==3.5.0
proto-plus==1.22.3
protobuf==4.24.4
psutil==5.9.5
pulsar-client==3.3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.19.0
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
Pygments==2.16.1
PyJWT==2.8.0
pymongo==4.6.0
pyparsing==3.1.1
PyPika==0.48.9
PySocks==1.7.1
pysrt==1.1.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
readchar==4.0.5
regex==2023.10.3
replicate==0.15.4
reportlab==4.0.6
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.3.1
scipy==1.11.3
selenium==4.14.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.21
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.9.26
tiktoken==0.4.0
tokenizers==0.14.1
tokentrim==0.1.10
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.11.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.6
uvicorn==0.23.2
uvloop==0.17.0
virtualenv==20.24.5
watchfiles==0.20.0
wcwidth==0.2.8
websockets==11.0.3
wget==3.2
wsproto==1.2.0
yarl==1.9.2
yaspin==3.0.1
yt-dlp==2023.9.24
zipp==3.17.0
